refactor: position 관련 패키지 분리
