+++ src/main/java/coursereview/springchess/domain/chesspiece/ChessPiece.java
@@ -1,6 +1,6 @@
 package coursereview.springchess.domain.chesspiece;
 
-import coursereview.springchess.domain.ChessPosition;
+import coursereview.springchess.domain.position.ChessPosition;
 
 public interface ChessPiece {
     boolean checkRule(ChessPosition source, ChessPosition target, boolean isEnemyOnTarget);
