+++ src/main/java/coursereview/springchess/domain/position/ChessPosition.java
@@ -1,4 +1,4 @@
-package coursereview.springchess.domain;
+package coursereview.springchess.domain.position;
 
 public enum ChessPosition {
     A1("a1"), A2("a2"), A3("a3"), A4("a4"), A5("a5"), A6("a6"), A7("a7"), A8("a8"),
